add quadratic funding logic
