+++ backend/app/engine/projects/__init__.py
@@ -1,9 +1,9 @@
 from dataclasses import dataclass, field
 
 from app.engine.projects.rewards import ProjectRewards
-from app.engine.projects.rewards.default import DefaultProjectRewards
+from app.engine.projects.rewards.quadratic_funding import QuadraticFundingProjectRewards
 
 
 @dataclass
 class ProjectSettings:
-    rewards: ProjectRewards = field(default_factory=DefaultProjectRewards)
+    rewards: ProjectRewards = field(default_factory=QuadraticFundingProjectRewards)
